requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
PyMySQL